requirements.txt
opencv-python
torch
numpy
pillow
plotly-express
matplotlib
kaleido
torchvision
numpy
opencv-python-headless
matplotlib
pyyaml
tqdm
easydict==1.9.0
scikit-image
scikit-learn==0.24.2
tensorflow
joblib
pandas
albumentations==0.5.2
hydra-core==1.1.0
pytorch-lightning==1.2.9
tabulate
kornia==0.5.0
webdataset
packaging
wldhx.yadisk-direct
altair<5